+++ Sources/HodEngine/Game/Components/TextureRendererComponent.cpp
@@ -103,7 +103,7 @@ namespace hod
 				std::shared_ptr<Node2dComponent> node2dComponent = entity->GetComponent<Node2dComponent>().lock();
 				if (node2dComponent != nullptr)
 				{
-					renderQueue.PushRenderCommand(new renderer::RenderCommandMesh(_vertices.data(), _vertices.size(), sizeof(renderer::P2fT2f), _indices.data(), _indices.size(), node2dComponent->GetWorldMatrix(), _materialInstance));
+					renderQueue.PushRenderCommand(new renderer::RenderCommandMesh(_vertices.data(), (uint32_t)_vertices.size(), sizeof(renderer::P2fT2f), _indices.data(), (uint32_t)_indices.size(), node2dComponent->GetWorldMatrix(), _materialInstance));
 				}
 			}
 		}
@@ -126,7 +126,7 @@ namespace hod
 					};
 
 					const renderer::Material* pickingMaterial = renderer::MaterialManager::GetInstance()->GetBuiltinMaterial(renderer::MaterialManager::BuiltinMaterial::P2fC4f_Unlit_Fill_Triangle);
-					renderQueue.PushRenderCommand(new renderer::RenderCommandMesh(vertices.data(), vertices.size(), sizeof(renderer::P2fC4f), _indices.data(), _indices.size(), node2dComponent->GetWorldMatrix(), pickingMaterial->GetDefaultInstance()));
+					renderQueue.PushRenderCommand(new renderer::RenderCommandMesh(vertices.data(), (uint32_t)vertices.size(), sizeof(renderer::P2fC4f), _indices.data(), (uint32_t)_indices.size(), node2dComponent->GetWorldMatrix(), pickingMaterial->GetDefaultInstance()));
 				}
 			}
 		}
